Move diagnostic interface to interface module
