+++ core/esmf-aspect-meta-model-interface/src/main/java/org/eclipse/esmf/Diagnostic.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright (c) 2026 Robert Bosch Manufacturing Solutions GmbH, Germany. All rights reserved.
+ */
+
+package org.eclipse.esmf;
+
+/**
+ * Generic diagnostic interface representing a problem found during document processing.
+ *
+ * @param <C> the concrete code type used to classify this diagnostic
+ */
+public interface Diagnostic<C extends Diagnostic.Code> {
+   C code();
+
+   String message();
+
+   Severity severity();
+
+   default boolean hasLocation() {
+      return false;
+   }
+
+   /**
+    * Identifies the kind of problem.
+    */
+   interface Code {
+      String code();
+
+      String description();
+   }
+
+   enum Severity {
+      ERROR,
+      WARNING,
+      INFO,
+      HINT
+   }
+}
