+++ core/esmf-aspect-meta-model-interface/src/main/java/org/eclipse/esmf/TurtleDiagnostic.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright (c) 2026 Robert Bosch Manufacturing Solutions GmbH, Germany. All rights reserved.
+ */
+
+package org.eclipse.esmf;
+
+public interface TurtleDiagnostic extends Diagnostic<Diagnostic.Code> {
+   enum TurtleCode implements Diagnostic.Code {
+      E0000( "No more info available" ),
+      E0001( "Could not load document" ),
+      E0002( "Document validation failed" ),
+      E0003( "Syntax error" ),
+      E0004( "Missing Token" );
+
+      private final String description;
+
+      TurtleCode( final String description ) {
+         this.description = description;
+      }
+
+      @Override
+      public String code() {
+         return name();
+      }
+
+      @Override
+      public String description() {
+         return description;
+      }
+   }
+}
