+++ core/esmf-tree-sitter-turtle/src/main/java/org/eclipse/esmf/treesitterturtle/TurtleSyntaxTree.java
@@ -22,6 +22,9 @@ import java.util.function.Supplier;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
+import org.eclipse.esmf.Diagnostic;
+import org.eclipse.esmf.TurtleDiagnostic;
+
 import org.jspecify.annotations.Nullable;
 import org.treesitter.TSNode;
 import org.treesitter.TSTree;
@@ -86,14 +89,9 @@ public class TurtleSyntaxTree {
       }
    }
 
-   public enum ErrorType {
-      SYNTAX_ERROR,
-      MISSING_TOKEN
-   }
-
    public record Error(
          String type,
-         ErrorType errorType,
+         Diagnostic.Code errorType,
          Location location
    ) implements Node {
       @Override
@@ -199,9 +197,9 @@ public class TurtleSyntaxTree {
             inputNode.getEndPoint().getRow(),
             inputNode.getEndPoint().getColumn() );
       if ( inputNode.isError() ) {
-         return new Error( inputNode.getType(), ErrorType.SYNTAX_ERROR, location );
+         return new Error( inputNode.getType(), TurtleDiagnostic.TurtleCode.E0003, location );
       } else if ( inputNode.isMissing() ) {
-         return new Error( inputNode.getType(), ErrorType.MISSING_TOKEN, location );
+         return new Error( inputNode.getType(), TurtleDiagnostic.TurtleCode.E0004, location );
       }
       final String token = tokenProvider.apply( location );
       final List<Node> children = IntStream.range( 0, inputNode.getChildCount() )
