+++ core/esmf-turtle-language-server/pom.xml
@@ -30,6 +30,10 @@
    <dependencies>
       <dependency>
          <groupId>org.eclipse.esmf</groupId>
+         <artifactId>esmf-aspect-meta-model-interface</artifactId>
+      </dependency>
+      <dependency>
+         <groupId>org.eclipse.esmf</groupId>
          <artifactId>esmf-aspect-meta-model-java</artifactId>
       </dependency>
       <dependency>
