+++ core/esmf-turtle-language-server/src/main/java/org/eclipse/esmf/turtle/languageserver/diagnostic/DiagnosticReport.java
@@ -15,6 +15,8 @@ package org.eclipse.esmf.turtle.languageserver.diagnostic;
 
 import java.util.List;
 
+import org.eclipse.esmf.TurtleDiagnostic;
+
 import com.google.common.collect.Streams;
 
 public record DiagnosticReport(
@@ -22,7 +24,7 @@ public record DiagnosticReport(
 ) {
    public static final DiagnosticReport EMPTY = new DiagnosticReport( List.of() );
 
-   public DiagnosticReport( final TurtleDiagnostic diagnostic ) {
+   public DiagnosticReport( final org.eclipse.esmf.TurtleDiagnostic diagnostic ) {
       this( List.of( diagnostic ) );
    }
